+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - New #103: Add class for tag `Form` and method `Html::form()` (vjik)
 - New #109: Add class for tag `Datalist` and method `Html::datalist()` (vjik)
 - New #109: Add specialized class for input tag with type `Range` and methods `Html::range()`, `Input::range()` (vjik)
+- New #113: Add class for tag `Legend`, class for tag `Fieldset`, methods `Html::legend()` and `Html::fieldset()` (vjik)
 - New #111: Add widget `ButtonGroup` (vjik)
 - New #111: Add method `Tag::unionAttributes()` that available for all tags (vjik)
 - Enh #106: Add option groups support to method `Select::optionsData()` (vjik)